Revert "Enum Pet"
