+++ AgendamentoBanhoETosa/Data/AppDbContext.cs
@@ -31,23 +31,6 @@ namespace AgendamentoBanhoETosa.Data
         public DbSet<Pet> Pets { get; set; }
         public DbSet<Servico> Servicos { get; set; }
 
-        // Configuração dos enums para inteiros
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            // Configura os enums TipoPet, RacaGato e RacaCachorro para serem armazenados como inteiros
-            modelBuilder.Entity<Pet>()
-                .Property(p => p.Tipo)
-                .HasConversion<int>();
 
-            modelBuilder.Entity<Pet>()
-                .Property(p => p.RacaGato)
-                .HasConversion<int>();
-
-            modelBuilder.Entity<Pet>()
-                .Property(p => p.RacaCachorro)
-                .HasConversion<int>();
-
-            base.OnModelCreating(modelBuilder);
-        }
     }
 }
