+++ AgendamentoBanhoETosa/Model/Cliente.cs
@@ -1,0 +1,11 @@
+﻿namespace AgendamentoBanhoETosa.Model
+{
+    public class Cliente
+    {
+        public int Id { get; set; }
+        public string Nome { get; set; }  
+        public string Telefone {  get; set; }
+        public string Endereco { get; set; }
+
+    }
+}
