--- AgendamentoBanhoETosa/Model/Entities/Cliente.cs
@@ -1,11 +1,0 @@
-namespace AgendamentoBanhoETosa.Model
-{
-    public class Cliente
-    {
-        public int Id { get; set; }
-        public string Nome { get; set; }  
-        public string Telefone {  get; set; }
-        public string Endereco { get; set; }
-
-    }
-}
