--- AgendamentoBanhoETosa/Model/Enums/ERacaCachorro.cs
@@ -1,115 +1,0 @@
-namespace AgendamentoBanhoETosa.Model.Enums
-{
-    public enum RacaCachorro
-    {
-        Affenpinscher,
-        AfghanHound,
-        AiredaleTerrier,
-        AkitaAmericano,
-        AmericanBully,
-        AmericanStaffordshireTerrier,
-        Basenji,
-        BassetHound,
-        Beagle,
-        BeardedCollie,
-        Bernese,
-        BichonHavanes,
-        BiewerTerrier,
-        Bloodhound,
-        BlueHeeler,
-        Boerboel,
-        BorderCollie,
-        BostonTerrier,
-        Boxer,
-        BracoAlemao,
-        Bulldog,
-        BulldogAlemao,
-        BulldogAmericano,
-        BulldogFrances,
-        BulldogIngles,
-        Bullmastiff,
-        BullTerrier,
-        CairnTerrier,
-        CaneCorso,
-        CavalierKingCharlesSpaniel,
-        Chihuahua,
-        ChowChow,
-        CockerSpaniel,
-        CockerSpanielAmericano,
-        CockerSpanielIngles,
-        Collie,
-        Corgi,
-        Danchshund,
-        Dalmata,
-        Doberman,
-        DogueAlemao,
-        DogueDeBordeaux,
-        FilaBrasileiro,
-        FlatCoatedRetriever,
-        FoxPaulistinha,
-        FoxTerrierDePeloDuro,
-        FoxTerrierDePeloLiso,
-        GoldenRetriever,
-        Greyhound,
-        HuskySiberiano,
-        JackRussellTerrier,
-        Keeshond,
-        Komondor,
-        Kuvasz,
-        LabradorRetriever,
-        LeaoDaRodesia,
-        LhasaApso,
-        LuluDaPomerania,
-        MalamuteDoAlasca,
-        Maltes,
-        Mastiff,
-        MastimEspanhol,
-        MastimNapolitano,
-        MastimTibetano,
-        NorfolkTerrier,
-        OldEnglishSheepdog,
-        PastorAlemao,
-        PastorAustraliano,
-        PastorBelga,
-        PastorBelgaGroenendael,
-        PastorBelgaMalinois,
-        PastorDeBeauce,
-        PastorDeShetland,
-        PastorDoCaucaso,
-        PastorMaremanoAbruzes,
-        PastorSuico,
-        Pequines,
-        Perdigueiro,
-        Pinscher,
-        Pitbull,
-        PointerIngles,
-        Poodle,
-        Pug,
-        Puli,
-        Rhodesia,
-        Rottweiler,
-        Saluki,
-        Samoieda,
-        SaoBernardo,
-        Schipperke,
-        Schnauzer,
-        SchnauzerGigante,
-        ScottishTerrier,
-        SetterIrlandes,
-        SharPei,
-        ShibaInu,
-        ShihTzu,
-        SpanielBretao,
-        SpitzJapones,
-        TerraNova,
-        TerrierTibetano,
-        Tervuren,
-        TosaInu,
-        ViraLata,
-        Weimaraner,
-        WelshCorgiCardigan,
-        WestHighlandTerrier,
-        Whippet,
-        YorkshireTerrier
-    }
-}
