--- AgendamentoBanhoETosa/Model/Enums/ERacaGato.cs
@@ -1,17 +1,0 @@
-namespace AgendamentoBanhoETosa.Model.Enums
-{
-    public enum RacaGato
-    {
-        Abissinio,
-        Angora,
-        Birmanes,
-        Bengal,
-        MaineCoon,
-        Persa,
-        Ragdoll,
-        ScottishFold,
-        Siames,
-        Sphynx,
-        ViraLata
-    }
-}
