--- AgendamentoBanhoETosa/Model/Enums/ETipoPet.cs
@@ -1,8 +1,0 @@
-namespace AgendamentoBanhoETosa.Model.Enums
-{
-    public enum TipoPet
-    {
-        Gato,
-        Cachorro
-    }
-}
