+++ AgendamentoBanhoETosa/appsettings.json
@@ -8,8 +8,5 @@
   "AllowedHosts": "*",
   "ConnectionStrings": {
     "AivenDB": "Host=pg-banhoetosa-josueramso-0611.e.aivencloud.com;Port=28086;Database=defaultdb;Username=avnadmin;;SslMode=Require"
-    //criei um novo banco no aiven so pra testar essa parte do enum 
-    //"Host=pg-27ef8c70-a0db.e.aivencloud.com;Port=16722;Database=defaultdb;Username=avnadmin;;SslMode=Require"
-
   }
 }
